Added error interceptor
